+++ src/clj_kondo/impl/config.clj
@@ -42,6 +42,7 @@
               :redefined-var {:level :warning}
               :redundant-declare {:level :warning}
               :var-same-name-except-case {:level :warning}
+              :constant-condition {:level :warning}
               :unreachable-code {:level :warning}
               :datalog-syntax {:level :error}
               :unbound-destructuring-default {:level :warning}
@@ -180,7 +181,6 @@
               :unused-alias {:level :off}
               :alias-same-as-ns {:level :off}
               :self-requiring-namespace {:level :warning}
-              :condition-always-true {:level :off}
               :underscore-in-namespace {:level :warning}
               :multiple-async-in-deftest {:level :warning}
               :java-static-field-call {:level :error}